+++ app/src/app/components/project/section/element/details.component.tsx
@@ -42,7 +42,6 @@ const ElementValuesViewer = ({ element, formModels, sectionId }: { element: ISec
             sectionId={sectionId}
             label={formModel.label}
             fieldName={formModel.fieldName}
-            value={element[formModel.fieldName]}
           />)
       }