+++ app/src/app/models/section/section.declarations.ts
@@ -29,6 +29,7 @@ export interface ISectionViewSettings {
   isHiddenInMenu?: boolean
   table?: {
     formElesVisibility?: { [key: string]: boolean }
+    sorting?: [string, 'asc' | 'desc' ] | null
   }
 }